+++ BetterRNG/BetterRNG.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Xna.Framework.Input;
@@ -106,7 +105,7 @@ namespace BetterRNG
             //Generate a good set of new random numbers to choose from for daily luck every morning.
             RandomFloats.FillFloats();
 
-            if(ModConfig.EnableDailyLuckOverride)
+            if (ModConfig.EnableDailyLuckOverride)
                 Game1.dailyLuck = RandomFloats.Random() / 10;
 
             if (ModConfig.EnableWeatherOverride)
@@ -132,228 +131,6 @@ namespace BetterRNG
                 ModConfig = ModConfig.ReloadConfig();
                 Log.Success("Config Reloaded for " + GetType().Name);
             }
-        }
-    }
-
-    public class RngConfig : Config
-    {
-        public bool EnableDailyLuckOverride { get; set; }
-        public bool EnableWeatherOverride { get; set; }
-        public List<string[]> Info { get; set; }
-        public int SunnyChance { get; set; }
-        public int CloudySnowyChance { get; set; }
-        public int RainyChance { get; set; }
-        public int StormyChance { get; set; }
-        public int HarshSnowyChance { get; set; }
-
-        public bool EnableFishingTreasureOverride { get; set; }
-        public float FishingTreasureChance { get; set; }
-        public bool EnableFishingStuffOverride { get; set; }
-
-        public override T GenerateDefaultConfig<T>()
-        {
-            EnableDailyLuckOverride = false;
-            EnableWeatherOverride = false;
-            Info = new List<string[]> { "The weather chances are whole numbers as percentages.|They can add up to be any number. 60 = 60% or 0.60, but you must type 60|The fishing things are not done yet and literally do nothing. Do not bother changing their values.".Split('|') };
-            SunnyChance = 60;
-            CloudySnowyChance = 15;
-            RainyChance = 15;
-            StormyChance = 5;
-            HarshSnowyChance = 5;
-
-            EnableFishingTreasureOverride = false;
-            FishingTreasureChance = 1 / 16f;
-            EnableFishingStuffOverride = false;
-            return this as T;
-        }
-    }
-
-
-
-
-
-
-
-    public static class Extensions
-    {
-        public static float[] DynamicDowncast(this Byte[] bytes)
-        {
-            float[] f = new float[bytes.Length / 4];
-            for (int i = 0; i < f.Length; i++)
-            {
-                f[i] = BitConverter.ToSingle(bytes, i == 0 ? 0 : (i * 4) - 1);
-            }
-            return f;
-        }
-
-        public static void FillFloats(this float[] floats)
-        {
-            for (int i = 0; i < floats.Length; i++)
-                floats[i] = BetterRng.Twister.Next(-100, 100) / 100f;
-        }
-
-        public static T Random<T>(this IEnumerable<T> enumerable)
-        {
-            if (enumerable == null)
-            {
-                throw new ArgumentNullException(nameof(enumerable));
-            }
-
-            var list = enumerable as IList<T> ?? enumerable.ToList();
-            return list.Count == 0 ? default(T) : list[BetterRng.Twister.Next(0, list.Count)];
-        }
-
-        public static float Abs(this float f)
-        {
-            return Math.Abs(f);
-        }
-
-        public static T Choose<T>(this List<T> list) where T : IWeighted
-        {
-            return Weighted.Choose(list);
-        }
-
-        public static T ChooseRange<T>(this List<T> list, Int32 min, Int32 max) where T : IWeighted
-        {
-            return Weighted.ChooseRange(list, min, max);
-        }
-
-        public static List<T> TempZero<T>(this List<T> list, int[] whichToZero) where T : class, IWeighted
-        {
-            T[] ls = new T[list.Count];
-            list.CopyTo(ls);
-
-            for (int index = 0; index < ls.Length; index++)
-            {
-                if (!whichToZero.Contains(index))
-                    continue;
-
-                T t = ls.ElementAtOrDefault(index);
-                if (t == null)
-                    continue;
-
-                t.Weight = 0;
-            }
-
-            return ls.ToList();
-        }
-
-        public static List<T> TempMult<T>(this List<T> list, float mult, int beginAt = 0) where T : class, IWeighted
-        {
-            T[] ls = new T[list.Count];
-            list.CopyTo(ls);
-
-            for (int index = beginAt; index < ls.Length; index++)
-            {
-                T t = ls.ElementAtOrDefault(index);
-                if (t == null)
-                    continue;
-
-                if (t.Weight == 0)
-                    continue;
-
-                t.Weight = mult > 0.5f ? (int)Math.Round(t.Weight * mult) : t.Weight;
-                mult *= 0.8f;
-            }
-
-            return ls.ToList();
-        }
-
-        public static int[] FishingZeroes(this Farmer player)
-        {
-            if (player == null)
-                player = BetterRng.Player;
-
-            return Enumerable.Range(0, player.FishingLevel - 3).ToArray();
-        }
-    }
-
-    public static class Weighted
-    {
-        public static T Choose<T>(List<T> list) where T : IWeighted
-        {
-            if (list.Count == 0)
-            {
-                return default(T);
-            }
-
-            int totalweight = list.Sum(c => c.Weight);
-            int choice = BetterRng.Twister.Next(totalweight);
-            int sum = 0;
-
-            foreach (var obj in list)
-            {
-                for (float i = sum; i < obj.Weight + sum; i++)
-                {
-                    if (i >= choice)
-                    {
-                        return obj;
-                    }
-                }
-                sum += obj.Weight;
-            }
-
-            return list.First();
-        }
-
-        public static T ChooseRange<T>(List<T> list, Int32 min, Int32 max) where T : IWeighted
-        {
-            //Grab a random value right quick, make sure the list is working and all
-            T choice = Choose(list);
-
-            //4096 changes to pick a number from the range.
-            //If it takes longer than that something isn't right.
-            for (int i = 0; i < 4096; i++)
-            {
-                choice = Choose(list);
-                if (choice.Value is Int32)
-                {
-                    Int32 x = (Int32)choice.Value;
-                    if (x >= min && x <= max)
-                        return choice;
-                }
-            }
-
-            if (choice is WeightedGeneric<int>)
-            {
-                int w = choice.Weight;
-                choice = Activator.CreateInstance<T>();
-                choice.Value = min < max ? BetterRng.Twister.Next(min, max) : BetterRng.Twister.Next(Math.Max(min, max));
-                choice.Weight = w;
-                return choice;
-            }
-
-            throw new ArgumentOutOfRangeException(nameof(min) + nameof(max), "The requested value was not in the Min/Max range supplied.");
-        }
-    }
-
-    public interface IWeighted
-    {
-        int Weight { get; set; }
-        object Value { get; set; }
-    }
-
-    public class WeightedGeneric<T> : IWeighted
-    {
-        public object Value { get; set; }
-        public int Weight { get; set; }
-
-        public T TValue => (T)Value;
-
-        public WeightedGeneric()
-        {
-            //Nothing
-        }
-
-        public WeightedGeneric(int weight, T value)
-        {
-            Weight = weight;
-            Value = value;
-        }
-
-        public static WeightedGeneric<Ty> Create<Ty>(int weight, Ty value)
-        {
-            return new WeightedGeneric<Ty>(weight, value);
         }
     }
 }
