+++ BetterRNG/IWeighted.cs
@@ -1,0 +1,8 @@
+﻿namespace BetterRNG
+{
+    public interface IWeighted
+    {
+        int Weight { get; set; }
+        object Value { get; set; }
+    }
+}
