+++ FishingMod/FishingMod.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="FishConfig.cs" />
     <Compile Include="FishingMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SBobberBar.cs" />
