+++ HealthBars/HealthBars.cs
@@ -79,7 +79,7 @@ namespace HealthBars
                 screenLoc.X += size.X / 2 - ModConfig.BarWidth / 2.0f;
                 screenLoc.Y -= ModConfig.BarHeight;
 
-                var fill = monster.health / (float) monster.maxHealth;
+                var fill = monster.health / (float)monster.maxHealth;
 
                 batch.Draw(texBar, screenLoc + new Vector2(ModConfig.BarBorderWidth, ModConfig.BarBorderHeight), texBar.Bounds, Color.Lerp(ModConfig.LowHealthColor, ModConfig.HighHealthColor, fill), 0.0f, Vector2.Zero, new Vector2(fill, 1.0f), SpriteEffects.None, 0);
 
@@ -110,51 +110,6 @@ namespace HealthBars
                 ModConfig = ModConfig.ReloadConfig();
                 Log.Success("Config Reloaded for " + GetType().Name);
             }
-        }
-    }
-
-    public class HealthBarConfig : Config
-    {
-        public bool DisplayHealthWhenNotDamaged { get; set; }
-
-        public bool DisplayMaxHealthNumber { get; set; }
-        public bool DisplayCurrentHealthNumber { get; set; }
-
-        public bool DisplayTextBorder { get; set; }
-
-        public Color TextColor { get; set; }
-        public Color TextBorderColor { get; set; }
-
-        public Color LowHealthColor { get; set; }
-        public Color HighHealthColor { get; set; }
-
-        public int BarWidth { get; set; }
-        public int BarHeight { get; set; }
-
-        public int BarBorderWidth { get; set; }
-        public int BarBorderHeight { get; set; }
-
-        public override T GenerateDefaultConfig<T>()
-        {
-            DisplayHealthWhenNotDamaged = false;
-
-            DisplayMaxHealthNumber = true;
-            DisplayCurrentHealthNumber = true;
-
-            DisplayTextBorder = true;
-
-            TextColor = Color.White;
-            TextBorderColor = Color.Black;
-
-            LowHealthColor = Color.DarkRed;
-            HighHealthColor = Color.LimeGreen;
-
-            BarWidth = 90;
-            BarHeight = 15;
-
-            BarBorderWidth = 2;
-            BarBorderHeight = 2;
-            return this as T;
         }
     }
 }
