+++ HealthBars/HealthBars.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="HealthBarConfig.cs" />
     <Compile Include="HealthBars.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
