+++ MovementMod/MovementMod.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="MovementConfig.cs" />
     <Compile Include="MovementMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
