+++ RegenMod/RegenMod.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="RegenConfig.cs" />
     <Compile Include="RegenMod.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
